requirements.txt
Pillow==10.2.0
customtkinter==5.2.2 